% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scFPCDE-package.R
\docType{package}
\name{scFPCDE-package}
\alias{scFPCDE}
\alias{scFPCDE-package}
\title{scFPCDE: FPCA-Based Differential Expression Along Pseudotime}
\description{
scFPCDE fits smooth gene-expression trajectories with functional principal
component analysis and performs permutation-based differential-expression
tests along pseudotime.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NLM-DIR/scFPCDE}
  \item Report bugs at \url{https://github.com/NLM-DIR/scFPCDE/issues}
}

}
\author{
\strong{Maintainer}: Ricardo J. Lopez Candelaria \email{rlopez.research@gmail.com}

}
\keyword{internal}
