% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permp_utils.R
\name{permp}
\alias{permp}
\title{Reverse-Rank Permutation P-values}
\usage{
permp(stats0, stats.perm, method = c("naive", "revrank"))
}
\arguments{
\item{stats0}{Vector of observed statistics}

\item{stats.perm}{Vector of permutation statistics}

\item{method}{Method to compute p-values: "naive" or "revrank"}
}
\value{
Vector of p-values
}
\description{
Helper function to compute permutation p-values using either naive or reverse-rank method.
}
\keyword{internal}
