% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scFPCDE_gene_curves.R
\name{scFPCDE_gene_curves}
\alias{scFPCDE_gene_curves}
\title{Plot Smoothed and Observed Gene Expression Curves}
\usage{
scFPCDE_gene_curves(
  tt,
  yt,
  yt_fit,
  cell_cluster,
  subset = NULL,
  facet_genes = TRUE,
  nrow = 7,
  ncol = 10,
  legend_dot_size = 3
)
}
\arguments{
\item{tt}{Vector of pseudotime values (length = # of timepoints)}

\item{yt}{Matrix of original expression data (rows = timepoints, columns = genes)}

\item{yt_fit}{Matrix of fitted expression values (same dimensions as yt)}

\item{cell_cluster}{Vector of cluster assignments (length = ncol(yt))}

\item{subset}{Optional vector of gene names or indices to subset}

\item{facet_genes}{If TRUE, plots genes in separate facets (default = TRUE)}

\item{nrow}{Number of facet rows (if faceted)}

\item{ncol}{Number of facet columns (if faceted)}

\item{legend_dot_size}{Size of legend dots for cluster colors}
}
\value{
A ggplot object
}
\description{
Visualizes fitted gene expression trajectories from FPCA alongside observed values, colored by cluster.
}
