% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scFPCDE_make_cluster.R
\name{scFPCDE_make_cluster}
\alias{scFPCDE_make_cluster}
\title{Cross-Platform Cluster Creation Helper}
\usage{
scFPCDE_make_cluster(
  ncores = ncores <- min(2, max(1, parallel::detectCores() - 1))
)
}
\arguments{
\item{ncores}{Number of cores to use (default = detectCores() - 1)}
}
\value{
A cluster object from the parallel package
}
\description{
Creates a parallel cluster using `makeForkCluster()` on Unix-like systems and
`makeCluster()` on Windows for compatibility.
}
\keyword{internal}
