% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scFPCDE_tune_fpca.R
\name{scFPCDE_tune_fpca}
\alias{scFPCDE_tune_fpca}
\title{Tune FPCA Parameters via Generalized Cross-Validation (GCV)}
\usage{
scFPCDE_tune_fpca(
  yt,
  tt,
  L = 3,
  r_pen_range = exp(-5:1),
  nbasis_range = 1:5 * 20,
  topvarper = 0.1
)
}
\arguments{
\item{yt}{Gene expression matrix (timepoints × genes)}

\item{tt}{Time vector}

\item{L}{Number of harmonics to retain}

\item{r_pen_range}{Vector of smoothing penalties}

\item{nbasis_range}{Vector of number of basis functions}

\item{topvarper}{Proportion of top signal genes used for tuning}
}
\value{
A list with optimal parameters and full GCV results
}
\description{
Optimizes smoothing parameters for FPCA using GCV on top signal curves.
}
